Rename Cypress snapshots runner
