+++ frontend/test/__runner__/run_cypress_tests.js
@@ -1,7 +1,7 @@
 const { printBold } = require("./cypress-runner-utils");
 const runCypress = require("./cypress-runner-run-tests");
 const getVersion = require("./cypress-runner-get-version");
-const generateSnapshots = require("./generate-cypress-snapshots");
+const generateSnapshots = require("./cypress-runner-generate-snapshots");
 
 // Use require for BackendResource to run it after the mock afterAll has been set
 const BackendResource = require("./backend.js").BackendResource;
